+++ bsp/sysp/port/luat_gpio_sysp.c
@@ -6,7 +6,7 @@ int luat_gpio_setup(luat_gpio_t* gpio) {
     return 0;
 }
 
-int luat_gpio_set(int pin, uint8_t level) {
+int luat_gpio_set(int pin, int level) {
     return 0;
 }
